protocol/requirements.txt
#
#######  Pip module requirements  #######
####### for the AGILE DBus Server #######
#
dbus-python
paho-mqtt
pydash
queuelib
pytz
liccheck
python-dotenv